+++ src/problems.rs
@@ -1,0 +1,3 @@
+pub mod bridge_and_torch;
+pub mod humans_and_zombies;
+pub mod wolf_goat_cabbage;